requirements.txt
pytest
python-dotenv
typer
msal
alive-progress